Create Coursera lab with jupyter.py
Solution for Methods and Classes lab on coursera python crash course
The code below defines an Elevator class. The elevator has a current floor, it also has a top and a bottom floor that are the minimum and maximum floors it can go to. Fill in the blanks to make the elevator go through the floors requested.